refactor: move UserRepository to user subpackage
- repository 패키지 내에 user 서브패키지를 추가하여 구조 정리
- UserRepository 인터페이스 및 MemoryUserRepository 클래스 이동
- 관련된 클래스(LoginServiceImpl.java, RegisterServiceImpl.java)의 import 및 코드 수정
- 코드 구조를 명확하게 정리하여 가독성 및 유지보수성 향상
